Address comments, update testing
