+++ src/types.js
@@ -1073,3 +1073,63 @@ export interface ShapeInformationInterface {
   getGetter(): void | SupportedGraphQLGetters;
   getAbstractType(): typeof Value;
 }
+
+type ECMAScriptType =
+  | "void"
+  | "null"
+  | "boolean"
+  | "string"
+  | "symbol"
+  | "number"
+  | "object"
+  | "array"
+  | "function"
+  | "integral";
+
+type ShapeDescriptorCommon = {
+  jsType: ECMAScriptType,
+  graphQLType?: string,
+};
+
+export type ShapePropertyDescriptor = {
+  shape: ShapeDescriptor,
+  optional: boolean,
+};
+
+type ShapeDescriptorOfObject = ShapeDescriptorCommon & {
+  kind: "object",
+  properties: { [string]: void | ShapePropertyDescriptor },
+};
+
+type ShapeDescriptorOfArray = ShapeDescriptorCommon & {
+  kind: "array",
+  elementShape: void | ShapePropertyDescriptor,
+};
+
+type ShapeDescriptorOfLink = ShapeDescriptorCommon & {
+  kind: "link",
+  shapeName: string,
+};
+
+type ShapeDescriptorOfPrimitive = ShapeDescriptorCommon & {
+  kind: "scalar",
+};
+
+type ShapeDescriptorOfEnum = ShapeDescriptorCommon & {
+  kind: "enum",
+};
+
+export type ShapeDescriptorNonLink =
+  | ShapeDescriptorOfObject
+  | ShapeDescriptorOfArray
+  | ShapeDescriptorOfPrimitive
+  | ShapeDescriptorOfEnum;
+
+export type ShapeDescriptor = ShapeDescriptorNonLink | ShapeDescriptorOfLink;
+
+export type ShapeUniverse = { [string]: ShapeDescriptor };
+
+export type ArgModel = {
+  universe: ShapeUniverse,
+  arguments: { [string]: string },
+};
